fill crate with loot action
