+++ A3UC/addons/admin/macros_zen.hpp
@@ -1,0 +1,6 @@
+    #define ICON_CURATOR            QUOTE(\a3\modules_f_curator\data\portraitcurator_ca.paa)
+    #define ICON_GENERIC            QUOTE(\a3\modules_f_curator\data\portraitdiary_ca.paa)
+    
+    #define ICON_PLAYER             QUOTE(\a3\modules_f_curator\data\portraitobjectivemove_ca.paa) // Replace with a person icon
+    #define ICON_OBJECT             QUOTE(\a3\modules_f_curator\data\portraitobjectivemove_ca.paa) // Replace with a box icon
+    #define ICON_OBJECT_INVENTORY   QUOTE(\a3\modules_f_curator\data\portraitobjectivemove_ca.paa) // Replace with an ammo icon
