+++ A3UC/addons/core/functions/utility/fn_isStatic.sqf
@@ -3,13 +3,13 @@
         Silence
     
     Description:
-        Checks if provided argument is an object
+        Checks if provided argument is a static turret
     
     Params:
         _object <OBJECT>
     
     Usage:
-        [cursorObject] call A3UC_fnc_isObject;
+        [cursorObject] call A3UC_fnc_isStatic;
     
     Return:
         true/false <BOOL>
@@ -21,9 +21,8 @@ if (_object isEqualType objNull) then {_object = typeOf _object};
 if (_object isKindOf "CAManBase") exitWith {false}; // object is player/unit
 if (_object isKindOf "Logic") exitWith {false}; // object is logic/module
 if (_object isKindOf "VirtualMan_F") exitWith {false}; // object is playable logic
-if ([_object] call A3UC_fnc_isVehicle) exitWith {false}; // object is vehicle
 
 // Things we want to be "objects"
-if (isClass (configFile >> 'CfgVehicles' >> _object)) exitWith {true};
+if (_object isKindOf "StaticWeapon") exitWith {true};
 
 false;
