+++ A3UC/addons/core/functions/utility/fn_isVehicle.sqf
@@ -19,8 +19,9 @@ params ["_vehicle"];
 
 if (_vehicle isEqualType objNull) then {_vehicle = typeOf _vehicle};
 if (_vehicle isKindOf "CAManBase") exitWith {false}; // object is player/unit
-
-// TO-DO: Add a check for statics and exit with false
+if (_vehicle isKindOf "Logic") exitWith {false}; // object is logic/module
+if (_vehicle isKindOf "VirtualMan_F") exitWith {false}; // object is playable logic
+if ([_vehicle] call A3UC_fnc_isStatic) exitWith {false}; // vehicle is static
 
 if (isClass (configFile >> 'CfgVehicles' >> _vehicle) && {_vehicle isKindOf "AllVehicles"}) exitWith {true};
